fixed the "angel 7-high bug"
This was caused when the AI asked for two rangers & got one. The
servers legion-list tracked it properly, but it then notified
all of the clients that both rangers had been received.

When it came time to aquire an angel the bug occurs

Two secondary problems that still exist consider:
-should the AI ask for two recruits when there is only one?
-should the AI ask for an angel when it thinks the stack is 7 high?


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2210 c5679a2a-0e0e-0410-bba2-8642b794b8c1