More engagement result dialog design changes.
Now the dialog moves to the new engagement if either
- the engagement happens in the player's turn
- it is the first one after the player's turn

The idea behind this design is that the dialog content moves along with the player
when the player is in charge of the game tempo, but if the player is only passive
the engagements stop moving until the player takes control by either continuing
to play or by clicking the next button.

The "Next" and "Hide" have been split into two buttons, each with their own
distinct functionality.

The GUI layout happens only once, except for the panels with the chits. The
latter should be changed later on since it is the only thing stopping us from
showing the empty dialog.


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2508 c5679a2a-0e0e-0410-bba2-8642b794b8c1