Part 2 of a project restructure.
This time the sources have been split into "core" and "datatools", which are compiled separately. 

The build results are now generated independently but in the same target directory called "build", not mixed with the sources anymore.

JavaMake didn't like this approach, so it is disabled at the moment. This needs more investigation -- it seems javamake deletes the contents of the target folder every time.


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2557 c5679a2a-0e0e-0410-bba2-8642b794b8c1