Refactoring: extracted common code for showing legion infos.
A new JPanel-based class can now be used to render a LegionInfo. This is used by ShowLegion and MasterBoard now.

A side effect of the refactoring is that the ShowLegion dialog now uses the player colour, too. This can be avoided by passing the colour as extra parameter, but I thought it doesn't look too bad this way. Let's wait for comments :-)

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2569 c5679a2a-0e0e-0410-bba2-8642b794b8c1