Fixed up dragging code for legion flyouts.
There was still a little problem left: the dragging was jumpy on the second and consecutive moves since the last mouse position was still set as the last one of the last drag.

To fix this the listener had to be both a MouseListener and a MouseMovementListener, to create that a new class was introduced. This class can be used to make any java.awt.Component draggable via a static method.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2570 c5679a2a-0e0e-0410-bba2-8642b794b8c1