Changed management of the old font in Marker's paint code.
For some reason the old font was stored as a member, which is not only unusual, but also introduces potential side effects since the font set at the end of the paint code might not be the one the graphic context had on entry (as it should be). Changed to standard pattern: store old setting in local var at the beginning, put back at the end.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2578 c5679a2a-0e0e-0410-bba2-8642b794b8c1