As discussed on the mailing list.
This patch is intended simply to always act on Phases and
BattlePhases as members of the appropriate type. Because the values
of these type are all constants, I've left the new types as
subclasses of Constants. Also because the values are all constants,
the == comparison operator is both legal and recommended. There
should be no user-visible change as the result of this pack. In
particular, 1590644 is still not fixed.



git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2594 c5679a2a-0e0e-0410-bba2-8642b794b8c1