The creature details dialog is now available through the Caretaker's Stack view.
This is related to #1594688 -- previously there was the option to show the details of a Ranger directly from the Graphics menu, but no other creature could be chosen.

This menu entry has been removed now, instead a click with the primary button on a label in the caretaker's stack shows the dialog for the creature clicked upon. This is not the only place where it could make sense and clicking the primary button not necessarily the best choice since it has other functions in other spots, but at least it gives access to the dialog for all creatures.


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2607 c5679a2a-0e0e-0410-bba2-8642b794b8c1