Workaround for detached AutoInspector bug.
The auto inspector window used to be detached from the master board (i.e. not on top, not raised together, lacking the icon) if it was created during game initialization. Clemens Katzer submitted this patch to work around the problem without changing the whole initialization process (the waffly comment is from me).

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2616 c5679a2a-0e0e-0410-bba2-8642b794b8c1