Startup log window with cancel option.
Another patch from Clemens Katzer: a window that shows startup progress (players log in) and allows aborting the process in case someone doesn't appear.

I changed the patch submitted to the mailing list by adding a checkbox for automatic closing when the game starts, which is on by default. The rationale behind this is that the information in the window becomes irrelevant to most users once the game has started. We could add storing the checkbox state between sessions, though.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2619 c5679a2a-0e0e-0410-bba2-8642b794b8c1