Configured explicit code templates for the Colossus project.
This is done to add the logger automatically to a new class -- both the two imports (Logger and Level) and the constant Logger instance.

The drawback of this is that this code is added even if not used and that Eclipse doesn't distinguish between classes and interfaces for the imports, which means the interfaces will get the logging imports, too. That doesn't make sense, but a Ctrl+Shift+O will fix it.


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2775 c5679a2a-0e0e-0410-bba2-8642b794b8c1