Cleaned up some code with potential performance issues.
The old code used to create a String[10][5] every time a method was called, which was effectively a Map<String,int[4]> that maps a terrain name to four types of bonuses. Not only was that array created on a regular basis, it was then iterated through to match the name and if there is a match the numbers stored as strings would be parsed into integers.

The whole thing is replaced with a constant HashMap now, initialized in a static initializer block. This maps the terrain name to an instance of a simple inner class modeling the bonuses. Should be a lot better, although I must admit that I didn't do any proper load testing. But IMO the new code is nicer, too.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2780 c5679a2a-0e0e-0410-bba2-8642b794b8c1