Run Jacobe code beautifier on sources.
I retrieved the old Jacobe config for the Colossus source (retrieved from an old revision in SVN) and ran it on the current source. Not too many changes, so we have been reasonably good.

Here is the command used (run inside the "core" directory, with Jacobe and the config in "tmp"):

../tmp/jacobe -cfg=../tmp/colossus-jacobe.cfg -overwrite -nobackup src/main/java/ src/test/java/

The config file can be found here in revision 2374 as trunk/CVSROOT/colossus-jacobe.cfg


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2791 c5679a2a-0e0e-0410-bba2-8642b794b8c1