Added option to display legion value in LegionInfoPanel.
If this option is set and the value of the legion is known to the player, a label at the end of the panel shows the legion's value. It is used in the legion flyouts at the moment, it can be toggled only by changing the constructor calls in the code right now.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2849 c5679a2a-0e0e-0410-bba2-8642b794b8c1