Done is disabled at the start now.
When moving from enabling/disabling the Done button to doing it on the action I missed some distinction in the old code about enabling/disabling depending on whose phase it is, which I partly fixed with one of the last commits. But the Done action still started out enabled, now it will be disabled until it is the local player's turn. It should do less toggles now in total, although I suspect the performance impact is negligible.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2852 c5679a2a-0e0e-0410-bba2-8642b794b8c1