Some changes towards using Swing for drawing instead just Java2D.
A lot of the code does the drawing itself where Swing should be used to do it -- thus breaking features such as finding the component under the mouse (and thus tooltips) and probably quite a few of the performance optimizations Swing does. It also created a lot of code that is far too complicated.

I started setting null layouts on some panels and adding chits and markers into the component hierarchy. Things are a bit tricky, so in many places I just went for TODO markers, I don't want to change too much close to a release.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2877 c5679a2a-0e0e-0410-bba2-8642b794b8c1