Code as formatted by Eclipse.
Some whitespace fixes, a couple of changes to the linebreaks. Unfortunately there are two cases where Eclipse does a bad job:

 - operators get onto the next line, e.g. "a\\+ b" not "a +\\b"
 - it likes breaking in front of a method call, e.g.
    "(.., client\\.getOptions())" instead of "(..,\\client.getOptions())"
   
I think both make the code slightly harder to read, but as discussed on the mailing list we consider it a price worth paying for being able to run the code formatter more easily.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2895 c5679a2a-0e0e-0410-bba2-8642b794b8c1