Eclipse formatter and similar settings.
This commits the configuration for the Eclipse formatter as discussed on the mailing list and used for the last commit.

The changes to the preferences also include the following "Save Actions" (only available in Eclipse 3.3+):

 - format source
 - organize imports
 - add final to private fields if possible
 - remove unnecessary casts
 - remove trailing whitespace (somehow that is a separate feature to formatting)
 
All these changes will be applied on each save in Eclipse 3.3 or higher. This last part is an experimental change -- we will see how we go with it.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2896 c5679a2a-0e0e-0410-bba2-8642b794b8c1