Split frame-related code from BattleMap.
The old BattleMap class was a JPanel itself, but also maintained the code for the JFrame around it and the other parts of the battle board -- plus the usual sprinkles of game logic one can find everywhere. The frame-related stuff has been moved out to a new class "BattleBoard", halfway into turning the BattleMap into a more or less dump panel for displaying the map itself (and fixing the component hierarchy).

Some code is still not really in the right position and some methods are just delegated from BattleBoard into BattleMap, but the first major split has been done.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2897 c5679a2a-0e0e-0410-bba2-8642b794b8c1