Turned terrain handling into typesafe enum for hazard terrains.
The old Terrains interface just introduced static final Strings for the hazard terrains. It is now replaced with a new typesafe enum HazardTerrain that not only adds the typesafety but also encapsulates some features intrinsic to a terrain.

There is no equivalent for the master board terrains yet since there were no constants for them to start with. It should be introduced one day, though.

The new class is also in a new package that finally should start collecting all the code relevant for games as in rules as well as in state. This is to avoid some of the redundant code between client and server: they should both depend on the game package instead. It should also collect some of the code from the GUI classes which is really not about GUI but about game state.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2899 c5679a2a-0e0e-0410-bba2-8642b794b8c1