Moved the AIs out of the client package into one of their own.
Since the AIs still heavily depend on parts of the client's code, a lot of the methods in the client package have to be public now instead of package-private. Since package-private doesn't really add much structure and no real security unless packages are sealed, I wouldn't consider that a loss -- after all package-private is hardly used and thus potentially confusing, too.

In the long term the AIs really shouldn't depend on the client, but the newly established game package.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2902 c5679a2a-0e0e-0410-bba2-8642b794b8c1