Changed handling of logging levels when changing the debug option.
The old code gives deprecation warnings against JDK 1.6 -- not a real issue, but I don't think it did what was intended, since the global logger previously changed would include logging against the JDK and other libraries.

Changed to using the logger for the toplevel Colossus package. Also changed the loglevel for the off case to actually turning it off -- it seems to match the description of the command line option that seems to be calling it.

Not sure this code is used at all -- part of me wants to remove it, but let's be a bit defensive.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2930 c5679a2a-0e0e-0410-bba2-8642b794b8c1