Bugfix: calculation of bonuses is corrected.
When moving from the String constants for the terrains to a typesafe enum I missed some usages of the literal strings instead of the constants, which caused some calculations of skill and strike bonuses to be wrong.

I did a textual search for the relevant strings so I think it should be all fixed now.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2933 c5679a2a-0e0e-0410-bba2-8642b794b8c1