Player name in status screen doesn't change color on death anymore.
Previously the player name would turn red after a player died. This could be annoying since the status screen uses the color codes to denote the players killed by another player -- if e.g. a player killed "gr" there was no way to tell which player that used to be.

The behavior is changed to keep the player's color in his/hers label even after death. The way I changed it it also keeps the color for the active player, which seems consistent. The rest of the column still turns red or yellow if required.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2943 c5679a2a-0e0e-0410-bba2-8642b794b8c1