First draft of new models for variants and games.
Two new packages try to model the concepts of "variant" (static information like available creatures, terrains, the master and battle board layouts, etc) and "game" (an ongoing game with its state).

This is more or less in draft state, the classes are near empty and only loosely connected to the existing code. Some matching existing classes already extend the new ones if possible -- the plan is to move code up into these new classes, which should be inherited by both client and server side classes. In the long run we hopefully will be able to remove the client and server specific classes completely in favor of the new ones.

There is some JavaDoc attached (including package.html files) that try to explain the design.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2964 c5679a2a-0e0e-0410-bba2-8642b794b8c1