More generics.
Eclipse still isn't too helpful, but luckily most of it is just pasting "<String>" all over the place. In turn it means there is a real need to introduce more typesafety, particularly avoiding doing things this way or the other (e.g. sometimes a List<Creature> and sometimes a List<String> is used for recruiters, the latter being their names).

The whole client package is generified. A lot of other little cleanups, too.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2975 c5679a2a-0e0e-0410-bba2-8642b794b8c1