Some code now uses Player instances as player IDs, not the names anymore.
It is not all code yet but we got rid of a lot of String comparisons. I am not sure if I like the Player/PlayerState distinction as it is at the moment, but things should be a bit more solid now.

The lifecycle of the Player instances is still a bit dodgy and held together by the automatic creation via the static Map in the Player class. It'll have to stay that way until the clients get smarter about using that class.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2997 c5679a2a-0e0e-0410-bba2-8642b794b8c1