Removed duplicated for loop.
In [3016] an old for loop wasn't replaced but extended by a new one, thus causing an duplicated iteration (n squared instead of n), which in turn caused the return value to be false (objects couldn't be removed the second time around), which then got caught by the unit tests.

Removed the old line.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3021 c5679a2a-0e0e-0410-bba2-8642b794b8c1