The startup progress window will appear only after 1 second now.
Instead of popping the window up and then closing it straight away, the window will delay showing itself for a second. If all players are in by then and the automatic close is enabled, it will never even show itself. That means that for local games the window stays out of the way.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3025 c5679a2a-0e0e-0410-bba2-8642b794b8c1