Separated MasterBoard (the model class) from MasterBoard (the JPanel).
All the master board model information is now in net.sf.colossus.variant.MasterBoard, with the MasterHex class moved along as well.

This not only splits the two concerns, it also removes a lot of static methods/members as well as dependencies from the server package into the client (nothing in the server package depends on the client.MasterBoard class anymore).


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3033 c5679a2a-0e0e-0410-bba2-8642b794b8c1