Removed Player class.
In the current version it stored only a name and a static map from names to players. No members beside the name are really needed right now, so the indirection didn't really make sense but instead caused the conflict on the term "player" between the player independent of a game and the player as participant inside a game.

The class (or some equivalent) might be needed again one day, but not now. And introducing it when needed should be easy enough.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3043 c5679a2a-0e0e-0410-bba2-8642b794b8c1