Moved option handling out of the client.
The Options class now allows attaching listeners to any option, which the client uses now instead of the internal option triggers. All option management is not done in the Options class, the Client does not implement IOptions anymore but instead a method getOptions() to allow access to the Options instance. It still implements setOption(String, String) for IClient (needed for network protocol), but that is just delegated.

The new approach would allow the various GUI elements to attach their own listeners instead of having the Client instance dispatch all changes. For example both the caretaker window and the checkbox for it in the menu could attach a listener each to control their state, thus decoupling them more from the client (and vice versa). That's not yet done.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3046 c5679a2a-0e0e-0410-bba2-8642b794b8c1