Much less referencing of players by their name.
The PlayerState object is passed around a lot more, the references to the player's name are slowly reduced to either the network protocol, file handling or logging. The latter should probably better rely on the toString() method instead of getName() since toString() used in string concatenation is null-safe, getName() isn't.

This commit cascaded throughout the code again, but the theme should stay the same everywhere.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3066 c5679a2a-0e0e-0410-bba2-8642b794b8c1