Another attempt at fixing the "no slayer" problem.
The check I introduced earlier checked the slayer string coming in from the network against the null value, when in fact the null value gets mapped to the "null" string. The check is adjusted and so far tests look good, although I haven't explicitly reproduced the mutual kill situation yet.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3079 c5679a2a-0e0e-0410-bba2-8642b794b8c1