Replaced a lot of "String markerId" with "Legion legion".
Many parameters and some return values don't use the markerId string anymore, but instead the Legion instances themselves, which gives the usual benefits of:
- more typesafe
- less code (no need to ask Game or Client for more information)
- less lookups at runtime

I had to add some more null-value handling in the network layer since there are cases where a markerId could have been null and needed to be replaced with a null Legion. Not nice, but unfortunately it has to happen somewhere as long as passing nulls on the network is allowed. It was done one way or the other before, just not as obvious.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3089 c5679a2a-0e0e-0410-bba2-8642b794b8c1