Preparing to move PlayerServerSide.legions up the hierarchy.
Originally I intended to replace all occurrences of legions in PlayerServerSide with the already existing (and thus confusing) member in Player, but (as so often) I got sidetracked into fixing a couple of other things first to ease that migration. Enough things to justify a separate test&commit cycle.

Mostly this moves some code of PlayerServerSide up into Player, changes some types from the server-specific versions to the generic ones and replaces some old-school loops with enhanced for-loops.


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3111 c5679a2a-0e0e-0410-bba2-8642b794b8c1