More consolidation of Player.getLegions().
The server-side now uses the List in the base class, not its own copy anymore. Client-side still stores the legions in the Client class, but access is now only through the Player object and as List<Legion>, not List<String> anymore. PlayerClientSide.getLegions() overrides the base method with a redirect to the Client, the relevant methods of Player access getLegions() instead of direct access to the member (this.legions) for now.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3112 c5679a2a-0e0e-0410-bba2-8642b794b8c1