More unification of the Legion hierarchy.
Some more methods move up, others got a common abstract method declaration in the base class.

I'm hitting the abstract methods now since the differences between the client side with the split prediction code and the server side with the actual game rules is getting larger. They should be something that can be fixed, but for now it seems better to let things be and focus on other stuff. TODOs are added with hints for further changes.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3123 c5679a2a-0e0e-0410-bba2-8642b794b8c1