Replaced hexLabel Strings with MasterHex instances.
Another one of these monster commits, with changes cascading all through the code. Most (all?) occurrences of labels for master hexes as parameters or return values are now replaced with using the MasterHex class.

There are some spots now where we could probably avoid passing the terrain as separate String parameter since it can be retrieved from the MasterHex instance. These are marked as TODOs, since this commit is big enough as it is I didn't even try to change those.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3132 c5679a2a-0e0e-0410-bba2-8642b794b8c1