Added a little dialog displaying a summary of a player.
This is currently still a bit rough, but the dialog already displays the name, type, score and titan power of the player, then a list of all legions with position and contents (as far as known) and then a list of all creatures owned.

Layout certainly could be better and currently the only access path is a bit hidden: it requires clicking on the player's name in the status screen. But it works.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3159 c5679a2a-0e0e-0410-bba2-8642b794b8c1