Fixed BalanceTowersTest.
MasterHexes can not retrieved as static members since there is no guarantee that a variant is loaded at that time. Even loading a variant in a static block before won't necessarily work since another (or the same) variant might have been loaded between loading the test class and running it, in which case the MasterHex instances are recreated.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3161 c5679a2a-0e0e-0410-bba2-8642b794b8c1