Added an experimental layout approach for the split history.
Not sure if that's going to work out -- the layout should be all right, but I don't really know how to display the necessary information inside the cells yet. The problem is that between splitting a legion off and splitting that new legion again the contents will have changed (mustering, summoning, kills). If it would be just growing things would be easy, e.g. you could represent a legion that was originally just two centaurs and then mustered another centaur and two lions as:

 C C + C L L

with the letters representing the matching chits. But the kills and creatures summoned out are more tricky.

And there is also the question what to do with killed legions -- they should probably end in the timeline.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3167 c5679a2a-0e0e-0410-bba2-8642b794b8c1