Bugfix: refusing to summon an angel should work again.
Currently there is no message that a summon was cancelled, instead a doSummon(null,null.null) is send. This message is important, though -- it causes the battle to progress out of the summon phase. To support this we need to be able to resolve the null markerId to a null legion, which we didn't do before.

Funny that didn't turn up in stresstesting -- it seems the AIs never refuse to summon.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3172 c5679a2a-0e0e-0410-bba2-8642b794b8c1