Introduced MasterBoardTerrain class.
This is the equivalent to HazardTerrain, but for the master board. It's nowhere near complete (e.g. there are still some maps in TerrainRecruitLoader that should probably be members) and it is not used everywhere it should (plenty of places where the getId() String is used instead). But it is a start.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3174 c5679a2a-0e0e-0410-bba2-8642b794b8c1