Games using the old AI classnames can be loaded without error.
If a player's (or client's) type is set to something starting with the old AI package name (i.e. the client package name), it will now automatically be mapped to the new AI package name.

This means that loading old savegames now gives you the AIs used instead of falling back onto SimpleAI with an error message.


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3190 c5679a2a-0e0e-0410-bba2-8642b794b8c1