Refined SCT logging a bit.
Instead of logging the message on the network on FINER, the SCT now logs the calls in the callMethod(..) dispatcher -- including when it is exited. This way a crash in executing a certain method can be found.

The incoming network level strings will still be logged on FINEST, the outgoing ones remain logged unchanged.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3193 c5679a2a-0e0e-0410-bba2-8642b794b8c1