Fixed loading games (mostly).
There is still an issue left where the masterboard displays some legions with the offset normally used for splits/battles. Not sure what causes that yet, but the game seems playable.

The fix is rather patchy. The problem is that when loading game part of it is restoring the final state, the other is replaying the history of events. The way I solved it at the moment is to avoid all changes to state that already has been restored, but that is rather messy and fragile.

It might be better to do it the other way around: replay the history first, then validate and extend the final state (the replay would not give the battle-specific information for the critters, although I think everything else should be there). Ideally the whole game would be set up around those events and they would be just replayed as is and that's it.



git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3200 c5679a2a-0e0e-0410-bba2-8642b794b8c1