Getting rid of more casts.
Since the Legion subtypes are not going to be merged completely anytime soon, it seems much better to expose the requirement for a subtype as soon as possible. I didn't touch the IClient or AI interfaces yet, but below that a lot of code has been moved to use LegionClientSide explicitly instead of casting.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3202 c5679a2a-0e0e-0410-bba2-8642b794b8c1