Made GUIHex covariant with Hex class.
The GUIHex class is now generic over the Hex class, with GUIBattleHex being a GUIHex<BattleHex> and GUIMasterHex being a GUIHex<MasterHex>. That means it is ensured the right type of model hex is stored inside, and the separate accessor methods with the downcasts are not needed anymore.

Also added some TODO and removed an old workaround for a JDK 1.3 bug from the code (which we don't support anymore).

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3213 c5679a2a-0e0e-0410-bba2-8642b794b8c1