Added function to access BattleTerrainHazardWindow from ShowBattleMap.
The way it works is far from perfect: only the primary-click into the ShowBattleMap window disposes it now, any other mouse button pops up the hazard window. Since it cleaned up some code, is useful for debugging the latter and it shouldn't hurt normal use (assuming people hit the primary button by default) I commit it as is.

One odd thing, though: for some reason once both windows are there the ShowBattleMap has to be disposed first, which is not what I would expect. They are both non-modal and have the masterboard frame as parent, so they should be independent.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3214 c5679a2a-0e0e-0410-bba2-8642b794b8c1