Changed Archdevil mustering from "Jungle" to "Tombs".
"Jungle" doesn't exist in this variant, so far Colossus just didn't catch that and just silently assumed an Archdevil can be mustered anywhere. "Tombs" uses the Jungle color and locations, so I assume that was meant by the original author -- I didn't get a definite answer on the mailing list, but it seems a pretty good guess.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3241 c5679a2a-0e0e-0410-bba2-8642b794b8c1