Fixed invalid implementation of contains() on Chit.
The spec says that the contains() method operates on the local coordinate system, while the implementation in Chit used the parent's coordinate system. This is fixed by removing the override and adjusting the caller (it's used only once in MasterBoard).

I also flatten the hierarchy a bit, instead of Chit being a JPanel without LayoutManager it is now a JComponent (which should effectively be the same).

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3492 c5679a2a-0e0e-0410-bba2-8642b794b8c1