Added task to create a buildinfo file.
The file contains some information about the Subversion working copy (revision, repository, changes) and the timestamp for the build. It should be included in the JAR if the "dist" target is called (since the SVN check can take a little time it is skipped otherwise).

Testing on different machines is required.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3713 c5679a2a-0e0e-0410-bba2-8642b794b8c1