Replaced the previous version number with the revision number.
Since the old version was just a build timestamp that seems an improvement. We have more information we could display (e.g. the timestamp of the build), but I believe the revision number including the modification flag should be sufficient.

Note that the main JAR file currently only contains this information if it was created using the "dist" target (or the "buildInfo" was called explicitly before the "jar"). This is a minor performance improvement to build the main JAR without having to check the working copy for changes, but it doesn't really take that long on my machine.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3769 c5679a2a-0e0e-0410-bba2-8642b794b8c1