Fixed: 2675045 NPE during createSummonAngel after Negotiation + related cleanup
In server side, handleNegotiation called CreateSummonAngel without the check whether there
are Legions with Donors at all. Now it checks.
Also, changed the findLegionsWithAngels argument to Legion instead of MarkerId;

Client side SummonAngel dialog bailed out early if donors list empty, and
in this situation saveWindow was not set yet so saveWindow.save() caused the NPE.
Added nullguard, just in case :)

On client side, pulled the FindLegionsWithAngels up to client which now first 
checks if the set is empty, and only if not empty pass the Set to the GUI or the AI.


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3800 c5679a2a-0e0e-0410-bba2-8642b794b8c1