Changed entry sides constants into enum.
This turned out to be much more surgery than expected. But mostly that was due to some special code in various classes which is largely consolidated in the new enum class now. Dependencies should be clearer and some redundancies should be gone.

The enum should probably be a toplevel class in the variants package, but I won't do that move just now.

There are also some other candidates for further optimization, but the change is big enough as it is.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3815 c5679a2a-0e0e-0410-bba2-8642b794b8c1