Moved AngelSummoningStates and LegionTags enum into BattleServerSide.
They are mostly used there, in fact the other users are either test cases or GameServerSide, the latter only since it still is responsible for restoring a battle from a savegame. If that would be moved inside the BattleServerSide class, then the enums could be private (the use in the constructor could be removed).


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3826 c5679a2a-0e0e-0410-bba2-8642b794b8c1