Changed GUI code to use anonymous inner classes.
I saw the comment about the dislike of them, but not only would I consider the use of anonymous inner classes standard for Swing, it also has advantages such as:

 * code stays close to the definition of the matching GUI element
 * no massive switch statements in the event listener
 * no accidental attachment of listeners where they are not needed
 * no need for lookup in datastructures to find the matching source
 * sometimes no need to store things in members at all
 * all in all more direct code

It helps to think of it as a typed closure instead of just weird syntax. And in total the number of lines should be down quite a bit, not up.

I also removed some dispose code that didn't really do anything.

All in all a lot of changes in the GUI, which unfortunately has some chance of breaking things. I played a game and opened all windows I can think of and it seemed ok. 

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3854 c5679a2a-0e0e-0410-bba2-8642b794b8c1