Replaced a lot of String parameter/members/return values with BattleHexes.
Another supposedly small refactoring that turned major surgery. Most of the String parameters in the code should be gone now, although at some places I still took shortcuts as to not increase the size of this change even more.

I had some bugs along the usual lines (non-generified methods on collections, identity issues), but I tested the code quite a bit and feel reasonably confident it is good now.

The usual amount of TODOs and minor changes along with the bigger change.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3904 c5679a2a-0e0e-0410-bba2-8642b794b8c1