Added two classes for easier Exception handling and took it into use in three places.
The helper classes make it easier to show a dialog box with the stacktrace data; the dialog box will be suppressed if running headless or in stresstest.

Added showing of such a dialog if an exception occurs during saving the game, during processing a line received from client, and in general in the most outer catch block just inside "server select loop" which waits for input on Server side.


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3925 c5679a2a-0e0e-0410-bba2-8642b794b8c1