Moved Legion->Marker binding from LegionClientSide to MasterBoard.
Previously the Marker was set on the LegionClientSide, despite being a GUI element. Both Client and ClientGUI would take some responsibility for the maintenance, making it a bit of a tangled web.

As it turns out, only MasterBoard cares about Markers, so I moved the link from the member of LegionClientSide into a LinkedHashMap on MasterBoard, taking the z-order functionality of the previous list of markers there.

I removed related code from Client and ClientGUI, although I am not sure I got all that could be removed. Client lost its dependency on Marker, ClientGUI still has it, but that doesn't seem to bad.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3937 c5679a2a-0e0e-0410-bba2-8642b794b8c1