Moved canRecruit() logic from LegionClientSide into Client.
This time we had two clients for the method: Client and SimpleAI, but since Simp
leAI knows about Client in many ways it seemed good to move canRecruit() to Clie
nt to get rid of that last case where LegionClientSide calls Client.

The client member of LegionClientSide is gone now and so is the dependency on Client.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3946 c5679a2a-0e0e-0410-bba2-8642b794b8c1