Went through UCDetector warnings on the MasterBoard class.
Removed some unused methods and a constant, a lot of visibility changed to private.

I'm not bothering with changing to default or protected visibility at the moment since that IMO doesn't gain much and would be rather tedious (there are a lot of those, and some methods are meant to be public, so decisions need to be made).

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3949 c5679a2a-0e0e-0410-bba2-8642b794b8c1