Renamed ExceptionUtils to ErrorUtils; extended it to provide also a method for "non-Exception" caused errors.
This can can be used for e.g. some messages in SocketClientThread which currently use some ClientGUI methods for that, but the ClientGUI might not be there yet, or not at all; I think it's good idea to show serious errors also for e.g. AIs that do have the NullGUI, because this way one gets a clue why the game hangs.


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4077 c5679a2a-0e0e-0410-bba2-8642b794b8c1