Started working on an event model for the game.
There are event object for various versions of the "add creature" event. It is set up in a hierarchy, with base classes/interfaces for the basic game event, events involving a legion and event revealing creatures.

The events are currently only used as parameters when moving information between game, server/client and other components. They should be stored in History later.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4186 c5679a2a-0e0e-0410-bba2-8642b794b8c1