Replaced == with .equals().
Even for typesafe enums the == is not safe. Multiple classloaders and the object serialization can still get you.

It is possibly related to some WARNINGs I see regarding different results for the attacker skill when striking up a wall. I'm still running the test, but either way the .equals() seems the better idea.

git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4206 c5679a2a-0e0e-0410-bba2-8642b794b8c1